requirements.txt
ollama
kokoro
sounddevice
soundfile
openai-whisper 
sounddevice 
soundfile 
keyboard 
numpy
flask